move threads.py and filter exceptions
